+++ src/endpoint.rs
@@ -76,7 +76,7 @@ impl Endpoint {
     }
 
     /// Indicates if the transfer descriptor is active
-    pub fn is_primed(&self, usb: &ral::usb::Instance) -> bool {
+    pub fn is_primed(&self, usb: &ral::AnyUsbInstance) -> bool {
         (match self.address.direction() {
             UsbDirection::In => ral::read_reg!(ral::usb, usb, ENDPTSTAT, ETBR),
             UsbDirection::Out => ral::read_reg!(ral::usb, usb, ENDPTSTAT, ERBR),
@@ -100,7 +100,7 @@ impl Endpoint {
 
     /// Initialize the endpoint, should be called soon after it's assigned,
     /// or after transitioning out of configuration (reset the endpoint).
-    pub fn initialize(&mut self, usb: &ral::usb::Instance) {
+    pub fn initialize(&mut self, usb: &ral::AnyUsbInstance) {
         if self.address.index() != 0 {
             let endptctrl = endpoint_control::register(usb, self.address.index());
             match self.address.direction() {
@@ -125,14 +125,14 @@ impl Endpoint {
     }
 
     /// Indicates if this endpoint has received setup data
-    pub fn has_setup(&self, usb: &ral::usb::Instance) -> bool {
+    pub fn has_setup(&self, usb: &ral::AnyUsbInstance) -> bool {
         ral::read_reg!(ral::usb, usb, ENDPTSETUPSTAT) & (1 << self.address.index()) != 0
     }
 
     /// Read the setup buffer from this endpoint
     ///
     /// This is only meaningful for a control OUT endpoint.
-    pub fn read_setup(&mut self, usb: &ral::usb::Instance) -> u64 {
+    pub fn read_setup(&mut self, usb: &ral::AnyUsbInstance) -> u64 {
         // Reference manual isn't really clear on whe we should clear the ENDPTSETUPSTAT
         // bit...
         //
@@ -178,7 +178,7 @@ impl Endpoint {
     }
 
     /// Clear the complete bit for this endpoint
-    pub fn clear_complete(&mut self, usb: &ral::usb::Instance) {
+    pub fn clear_complete(&mut self, usb: &ral::AnyUsbInstance) {
         match self.address.direction() {
             UsbDirection::In => {
                 ral::write_reg!(ral::usb, usb, ENDPTCOMPLETE, ETCE: 1 << self.address.index())
@@ -193,7 +193,7 @@ impl Endpoint {
     ///
     /// Caller should check to see if there is an active transfer, or if the previous
     /// transfer resulted in an error or halt.
-    pub fn schedule_transfer(&mut self, usb: &ral::usb::Instance, size: usize) {
+    pub fn schedule_transfer(&mut self, usb: &ral::AnyUsbInstance, size: usize) {
         self.td.set_terminate();
         self.td.set_buffer(self.buffer.as_ptr_mut(), size);
         self.td.set_interrupt_on_complete(true);
@@ -216,7 +216,7 @@ impl Endpoint {
     }
 
     /// Stall or unstall the endpoint
-    pub fn set_stalled(&mut self, usb: &ral::usb::Instance, stall: bool) {
+    pub fn set_stalled(&mut self, usb: &ral::AnyUsbInstance, stall: bool) {
         let endptctrl = endpoint_control::register(usb, self.address.index());
 
         match self.address.direction() {
@@ -230,7 +230,7 @@ impl Endpoint {
     }
 
     /// Indicates if the endpoint is stalled
-    pub fn is_stalled(&self, usb: &ral::usb::Instance) -> bool {
+    pub fn is_stalled(&self, usb: &ral::AnyUsbInstance) -> bool {
         let endptctrl = endpoint_control::register(usb, self.address.index());
 
         match self.address.direction() {
@@ -242,7 +242,7 @@ impl Endpoint {
     /// Enable the endpoint
     ///
     /// This should be called only after the USB device has been configured.
-    pub fn enable(&mut self, usb: &ral::usb::Instance) {
+    pub fn enable(&mut self, usb: &ral::AnyUsbInstance) {
         // EP0 is always enabled
         if self.address.index() != 0 {
             let endptctrl = endpoint_control::register(usb, self.address.index());
@@ -260,7 +260,7 @@ impl Endpoint {
     /// Indicates if this endpoint is enabled
     ///
     /// Endpoint 0, the control endpoint, is always enabled.
-    pub fn is_enabled(&self, usb: &ral::usb::Instance) -> bool {
+    pub fn is_enabled(&self, usb: &ral::AnyUsbInstance) -> bool {
         if self.address.index() == 0 {
             return true;
         }
@@ -273,7 +273,7 @@ impl Endpoint {
     }
 
     /// Clear the NACK bit for this endpoint
-    pub fn clear_nack(&mut self, usb: &ral::usb::Instance) {
+    pub fn clear_nack(&mut self, usb: &ral::AnyUsbInstance) {
         match self.address.direction() {
             UsbDirection::In => {
                 ral::write_reg!(ral::usb, usb, ENDPTNAK, EPTN: 1 << self.address.index())