+++ src/log.rs
@@ -4,9 +4,8 @@
 
 macro_rules! trace {
     ($($args:tt)*) => {
-        #[cfg(feature = "defmt-03")]
+        #[cfg(feature = "defmt")]
         {
-            use defmt_03 as defmt;
             defmt::trace!($($args)*)
         }
     };
@@ -14,9 +13,8 @@ macro_rules! trace {
 
 macro_rules! debug {
     ($($args:tt)*) => {
-        #[cfg(feature = "defmt-03")]
+        #[cfg(feature = "defmt")]
         {
-            use defmt_03 as defmt;
             defmt::debug!($($args)*)
         }
     };
@@ -24,9 +22,8 @@ macro_rules! debug {
 
 macro_rules! info {
     ($($args:tt)*) => {
-        #[cfg(feature = "defmt-03")]
+        #[cfg(feature = "defmt")]
         {
-            use defmt_03 as defmt;
             defmt::info!($($args)*)
         }
     };
@@ -34,9 +31,8 @@ macro_rules! info {
 
 macro_rules! warn {
     ($($args:tt)*) => {
-        #[cfg(feature = "defmt-03")]
+        #[cfg(feature = "defmt")]
         {
-            use defmt_03 as defmt;
             defmt::warn!($($args)*)
         }
     };